Fix marquee scenelet
It might ultimately be better not to include this particular scenelet,
but for now it needs to be fixed because it was broken.  I also replaced
almost all of the Color constructor calls with predefined colors, to
avoid unforeseen breakage in the miniRT when I eventually get around to
changing Color to use floating point.